requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.0
sqlalchemy>=2.0.30
python-multipart>=0.0.9
reportlab>=4.2.0
requests>=2.32.0
numpy>=1.26.4
torch
torchvision
opencv-python
librosa
exifread
Pillow
matplotlib
